# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: REST api for simple route app
  version: ""
host: /
consumes:
  - application/json
  - application/xml
  - application/gob
produces:
  - application/json
  - application/xml
  - application/gob
paths:
  /api/admin/bots/:
    post:
      tags:
        - admin_service
      summary: push_bots admin_service
      description: |-
        push bots to parser service

        **Required security scopes for jwt**:
          * `admin`
      operationId: admin_service#push_bots
      parameters:
        - name: Authorization
          in: header
          description: JWT used for authentication
          required: true
          type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminServicePushBotsOKResponseBody'
            required:
              - sent_bots
              - saved_bots
              - usernames
        "400":
          description: Bad Request response.
          schema:
            type: string
        "401":
          description: Unauthorized response.
          schema:
            type: string
        "403":
          description: Forbidden response.
          schema:
            type: string
        "404":
          description: Not Found response.
          schema:
            type: string
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/AdminServicePushBotsInternalServerErrorResponseBody'
            required:
              - sent_bots
              - saved_bots
              - usernames
      schemes:
        - http
      security:
        - jwt_header_Authorization: []
  /api/admin/managers/:
    post:
      tags:
        - admin_service
      summary: add_manager admin_service
      description: |-
        admins could add drivers from main system

        **Required security scopes for jwt**:
          * `admin`
      operationId: admin_service#add_manager
      parameters:
        - name: Authorization
          in: header
          description: JWT used for authentication
          required: false
          type: string
        - name: add_manager_request_body
          in: body
          required: true
          schema:
            $ref: '#/definitions/AdminServiceAddManagerRequestBody'
            required:
              - login
              - password
      responses:
        "200":
          description: OK response.
        "400":
          description: Bad Request response.
          schema:
            type: string
        "401":
          description: Unauthorized response.
          schema:
            type: string
        "403":
          description: Forbidden response.
          schema:
            type: string
        "404":
          description: Not Found response.
          schema:
            type: string
      schemes:
        - http
      security:
        - jwt_header_Authorization: []
  /api/profile:
    get:
      tags:
        - auth_service
      summary: profile auth_service
      description: get user profile
      operationId: auth_service#profile
      parameters:
        - name: Authorization
          in: header
          description: JWT used for authentication
          required: true
          type: string
      responses:
        "200":
          description: OK response.
        "400":
          description: Bad Request response.
          schema:
            type: string
        "401":
          description: Unauthorized response.
          schema:
            type: string
        "404":
          description: Not Found response.
          schema:
            type: string
        "500":
          description: Internal Server Error response.
          schema:
            type: string
      schemes:
        - http
      security:
        - jwt_header_Authorization: []
...
definitions:
  AdminServiceAddManagerRequestBody:
    title: AdminServiceAddManagerRequestBody
    type: object
    properties:
      login:
        type: string
        example: Soluta quaerat eveniet dolor sit ex nihil.
      password:
        type: string
        example: jgb
        minLength: 4
    example:
      login: Cupiditate non quaerat dolore voluptas ipsa provident.
      password: ic2
    required:
      - login
      - password
  AdminServicePushBotsInternalServerErrorResponseBody:
    title: AdminServicePushBotsInternalServerErrorResponseBody
    type: object
    properties:
      saved_bots:
        type: integer
        description: количество ботов, которых сохранили в проксе
        example: 1161699258
        format: int32
      sent_bots:
        type: integer
        description: количество ботов, которых мы отправили
        example: 4163079019522061408
        format: int64
      usernames:
        type: array
        items:
          type: string
          example: Ex et aliquam tempore.
        description: имена ботов, которые мы сохранили
        example:
          - Quis dolorum enim qui officia est earum.
          - Beatae doloribus et facere similique.
          - Voluptates fugit repellendus debitis provident ipsa.
          - Deserunt quis omnis fuga eius eum natus.
    example:
      saved_bots: 613302728
      sent_bots: 6265530653151255847
      usernames:
        - Dolores laboriosam omnis esse nulla.
        - Consectetur est mollitia.
        - Natus tenetur tenetur ipsa.
    required:
      - sent_bots
      - saved_bots
      - usernames
...
  AdminServicePushBotsOKResponseBody:
    title: AdminServicePushBotsOKResponseBody
    type: object
    properties:
      saved_bots:
        type: integer
        description: количество ботов, которых сохранили в проксе
        example: 1908492623
        format: int32
      sent_bots:
        type: integer
        description: количество ботов, которых мы отправили
        example: 1343352821286468698
        format: int64
      usernames:
        type: array
        items:
          type: string
          example: Velit aut odit accusamus.
        description: имена ботов, которые мы сохранили
        example:
          - Ut non.
          - Nobis aliquid.
    example:
      saved_bots: 1819366484
      sent_bots: 8080549128743596134
      usernames:
        - Ducimus explicabo error id qui et aut.
        - Nesciunt ut qui sequi et unde eum.
    required:
      - sent_bots
      - saved_bots
      - usernames
...
securityDefinitions:
  basic_header_Authorization:
    type: basic
    description: |-
      Basic authentication used to authenticate security principal during signin

      **Security Scopes**:
        * `driver`: Read-only access
  jwt_header_Authorization:
    type: apiKey
    description: |-
      Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes "api:read" and "api:write".

      **Security Scopes**:
        * `driver`: Read-only access
        * `admin`: Read and write access
    name: Authorization
    in: header